slam-rs: resolve catalog sources once and split pure feed operations
